[#10] - Refactored database file and tests
