+++ .gitignore
@@ -3,3 +3,4 @@
 .idea
 .vscode
 /dist
+database.sqlite
